Refactored lib/auth/services/auth_user.dart -> lib/auth/services/auth_services/auth_user.dart
